# speed up lm_pd.fit