Hi @const-ae 

I had a look on  the `lm_pd.fit` function, since it seems this causes the speed limitations (it is called several times in the function `fit_parameters_loop`, especially in `lapply`). By itself, the function is very fast and I don't think we can speed it up much more (especially the optimization functions that are already written in C). 

I found some minor issues that make it slightly faster, maybe in the sum then it will run a bit faster (at least my `microbenchmark` checks showed some improvement in computation time). 

I was also thinking of parallelizing the `lapply` that uses the `lm_pd.fit`, but so far it seems making up the cluster takes more time (since every fitting by `lm_pd.fit` is quite fast), maybe chunking will help but I will have a look on this.