@const-ae 2021-01-12
Hi Thomas,

thanks for taking the time and digging into the issue :)

However, I took a brief look at the diff (https://github.com/const-ae/proDA/pull/8/files) and it is very hard to tell what has actually changed, because the indentation seems off. This makes it basically impossible review the PR, because I can't tell what are the relevant changes. Would it be possible to make a new PR where you only change the relevant lines?

Best,
Constantin