@tnaake 2021-01-12
Hi again,

sorry for the confusion with the coding style - hope it's more readable now. I mainly changed such that within the functions reused objects are not recalculated, with this we got slightly better performance (my idea is that when `pd_lm.fit` is run > 1000 times we get a better performance, otherwise the proposed changes are negligible. 


benchmark for `grad_fnc`    

    microbenchmark(
        fct1 = proDA:::grad_fnc(y, yo, X, Xm, Xo,
                     beta, sigma2, rho, zetastar, mu0, sigma20, df0, tau20,
                     location_prior_df, moderate_location, moderate_variance),
        fct2 = grad_fnc(y, yo, X, Xm, Xo,
              beta, sigma2, rho, zetastar, mu0, sigma20, df0, tau20,
              location_prior_df, moderate_location, moderate_variance)
    )

    ## Unit: microseconds
    ## expr  min   lq    mean median    uq     max neval cld
    ## fct1 34.6 36.2 142.750   36.7 37.75 10505.2   100   a
    ## fct2 34.4 35.4  36.992   36.2 37.20    70.6   100   a

benchmark for `dt.scaled`

    microbenchmark(
        sc1 = proDA:::dt.scaled(x, df), 
        sc2 = dt.scaled(x, df))

    ##Unit: microseconds
    ##expr min  lq  mean median  uq  max neval cld
    ##sc1 4.7 4.8 5.661   4.85 5.1 40.5   100   a
    ##sc2 4.4 4.5 4.788   4.60 4.8 11.4   100   a


benchmark for `objective_fnc`

    microbenchmark(
        obj1 = proDA:::objective_fnc(y, yo, X, Xm, Xo, beta, sigma2, rho, zetastar, mu0, 
                                 sigma20, df0, tau20, location_prior_df, moderate_location, 
                                 moderate_variance),
        obj2 = objective_fnc(y, yo, X, Xm, Xo, beta, sigma2, rho, zetastar, mu0, 
                         sigma20, df0, tau20, location_prior_df, moderate_location, 
                         moderate_variance)
    )
    
    ## Unit: microseconds
    ## expr  min    lq   mean median    uq    max neval cld
    ## obj1 11.8 12.55 97.466  13.90 16.95 8189.4   100   a
    ## obj2 11.7 12.20 15.389  12.75 14.75   41.1   100   a