@tnaake 2021-01-12
benchmark for `hess_fnc` *here actually the implemented version is faster, why?)

    microbenchmark(
        hess2 = hess_fnc(y, yo, X, Xm, Xo, beta, sigma2, rho, zetastar, mu0, 
                     sigma20, df0, tau20, location_prior_df, moderate_location, 
                     moderate_variance, beta_sel, p),
        hess1 = proDA:::hess_fnc(y, yo, X, Xm, Xo, beta, sigma2, rho, zetastar, mu0, 
                     sigma20, df0, tau20, location_prior_df, moderate_location, 
                     moderate_variance, beta_sel, p),
         times = 10000
    )

    ## Unit: microseconds
    ## expr  min    lq     mean median    uq     max neval cld
    ## hess2 87.4 104.1 129.0845  110.6 138.2 16117.3 10000   b
    ##hess1 83.0  97.5 119.2408  103.5 127.4  4272.9 10000  a 

microbenchmark for `pd_lm.fit` (started with a difference of +0.07 now it is faster, I didnt use proDA:::pd_lm.fit here but compared it to `pd_lm.fit <- proDA:::pd_lm.fit`, strangely these give slightly different times)

    microbenchmark(
        fit1 = pd_lm.fit_old(y, X, dropout_curve_position, dropout_curve_scale, location_prior_mean, location_prior_scale,
            variance_prior_scale, variance_prior_df, location_prior_df, method = "analytic_hessian", verbose = FALSE),
        fit2 = pd_lm.fit_new(y, X, dropout_curve_position, dropout_curve_scale, location_prior_mean, location_prior_scale,
                         variance_prior_scale, variance_prior_df, location_prior_df, method = "analytic_hessian", verbose = FALSE)
    )

    ##Unit: milliseconds
    ## expr    min      lq     mean  median      uq    max neval cld
    ## fit1 1.3728 1.58320 1.815087 1.68075 1.82435 4.2162   100   a
    ## fit2 1.3714 1.56735 1.801352 1.67975 1.90655 5.1906   100   a