@const-ae 2021-01-12
This is great. Thanks, I will give more detailed feedback on the code later.

I compiled a quick overview table of the performance improvements:

| function        | Median (old / new) | Mean (old / new) |
|-----------------|--------------------|------------------|
| objective_fnc() | 13.9 / 12.8 µs     | 97 / 12.8 µs     |
| grad_fnc()      | 36.7 / 36.2 µs     | 142.8 / 37.0 µs  |
| hess_fnc()      | 103.5 / 110.6 µs   | 119.2 / 129.1 µs |
| pd_lm.fit()     | 1.68 / 1.68 ms     | 1.81 / 1.80 ms   |
| dt.scaled()     | 4.9 / 4.6 µs       | 5.6 / 4.8 µs     |

1) Do you know where the big difference between the median and mean timings for the old code comes from? Could it be related to garbage collection? Do you see the same pattern with the `bench::mark()` function?

2) How come the performance improvements for the `objective_fnc`, `grad_fnc` does not translate into a faster `pd_lm.fit()`? Are those actually the bottlenecks for execution speed if you look at the flamegraph with `profvis`?